requirements.txt
Django==3.2
djangorestframework==3.12.4
flake8==3.9.1
isort==5.8.0
black==20.8b1
